Various code quality improvements
- Fix a few remaining `with` chains not found before
- Move eval handler into eval.go to reduce size of server.go
- Add util.Or convenience function
- Remove unused "loaded rules" logic
- Test runner API improvements before upcoming overhaul
