+++ bundle/regal/lsp/semantictokens/vars/imports/imports_test.rego
@@ -9,7 +9,8 @@ import data.regal.ast`
 
 	module := regal.parse_module("p.rego", policy)
 
-	tokens := imports.result with input.params as {"textDocument": {"uri": "file://p.rego"}}
+	tokens := imports.result
+		with input.params as {"textDocument": {"uri": "file://p.rego"}}
 		with input.regal as module.regal
 		with data.workspace.parsed["file://p.rego"] as module
 
@@ -27,7 +28,8 @@ import data.other.identifier as alias`
 
 	module := regal.parse_module("p.rego", policy)
 
-	tokens := imports.result with input.params as {"textDocument": {"uri": "file://p.rego"}}
+	tokens := imports.result
+		with input.params as {"textDocument": {"uri": "file://p.rego"}}
 		with input.regal as module.regal
 		with data.workspace.parsed["file://p.rego"] as module
